// frontend/src/api/generated/models/TopicMastery.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Mastery level per topic.
 */
export type TopicMastery = {
    topic: string;
    mastery_score: number;
    card_count: number;
    avg_ease_factor: number;
};